+++ changelog.md
@@ -1,11 +1,12 @@
 Codion Change Log
-==================
+=================
 
 ## 0.18.29
 ### is.codion.swing.common.ui
 - FilterTable.ScrollToAdded bug fixed, now stops scrolling after the first row.
 - FilterTable.Builder.surrendersFocusOnKeystroke() added.
 - DefaultFilterTableCellEditor now sets JComboBox.isTableCellEditor client property for combo boxes.
+- FilterTable bug fixed, updateUI() now updates cell editor UIs.
 ### is.codion.framework.db
 - EntityConnectionProvider.connection() inferred return type removed.
 - AbstractEntityConnectionProvider.validateConnection() replaced with validConnection(), connection() no longer final.